Moved ensure_rng to top level utils
